+++ src/__tests__/typedefs.test.js
@@ -94,11 +94,7 @@ struct UserActivitiesRequest {
     .filter(p => p.endsWith('.thrift'))
     .map(p => path.resolve(root, p))
     .forEach(p => {
-      let output = new ThriftFileConverter(
-        p,
-        name => name,
-        true
-      ).generateFlowFile();
+      let output = new ThriftFileConverter(p, true).generateFlowFile();
       let longIndex = output.indexOf('import Long');
       expect(longIndex).not.toBe(-1);
     });
@@ -118,11 +114,7 @@ typedef i64 (js.type = "Long") Points
     .filter(p => p.endsWith('.thrift'))
     .map(p => path.resolve(root, p))
     .forEach(p => {
-      let output = new ThriftFileConverter(
-        p,
-        name => name,
-        true
-      ).generateFlowFile();
+      let output = new ThriftFileConverter(p, true).generateFlowFile();
       let longIndex = output.indexOf('import Long');
       // Expected long definition but did not find one
       expect(longIndex).not.toBe(-1);