+++ src/main/index.js
@@ -4,30 +4,20 @@ import {ThriftFileConverter} from './convert';
 import commonPathPrefix from 'common-path-prefix';
 import path from 'path';
 
-type OptionsType = {
-  suffix?: string,
+type OptionsType = {|
   withSource?: boolean,
   commonPath?: string,
   outputDir?: string,
-};
+|};
 
 export default function convert(
   thriftPaths: Array<string>,
   options: OptionsType
 ) {
-  const {
-    suffix = 'Type',
-    withSource = false,
-    outputDir = 'flow-output',
-    commonPath,
-  } = options;
+  const {withSource = false, outputDir = 'flow-output', commonPath} = options;
   const allOutput = {};
   for (const thriftPath of thriftPaths) {
-    const converter = new ThriftFileConverter(
-      thriftPath,
-      name => name + suffix,
-      withSource
-    );
+    const converter = new ThriftFileConverter(thriftPath, withSource);
     converter
       .getImportAbsPaths()
       .filter(p => thriftPaths.indexOf(p) === -1)