+++ src/main/thrift-types.js
@@ -46,14 +46,9 @@ export class TypeConverter {
     Date: 'string',
   };
 
-  transformName: string => string;
   thriftAstDefinitions: Array<any>;
 
-  constructor(
-    transformName: string => string,
-    thriftAstDefinitions: Array<any>
-  ) {
-    this.transformName = transformName;
+  constructor(thriftAstDefinitions: Array<any>) {
     this.thriftAstDefinitions = thriftAstDefinitions;
   }
 
@@ -73,13 +68,13 @@ export class TypeConverter {
       this.mapType(t) ||
       this.annotation(t) ||
       TypeConverter.primitives[t.baseType] ||
-      this.transformName(id(t.name))
+      id(t.name)
     );
   };
 
   enumType = (thriftValueType: BaseType) => {
     if (this.isEnum(thriftValueType)) {
-      const name = this.transformName(thriftValueType.name);
+      const name = thriftValueType.name;
       // Enums are values, not types. To refer to the type,
       // we use $Values<...>.
       return `$Values<typeof ${name}>`;